requirements.txt
vk-api==11.9.9
python-dotenv==1.0.0
Flask==3.0.3
gunicorn==21.2.0